move generateAddPriceTable
