+++ app/helpers/contextCreators/generateAddPriceTable.js
