+++ app/pages/sections/catalogue-solutions/order-item/contextCreator.js
@@ -2,7 +2,7 @@ import { baseUrl } from '../../../../config';
 import { generateErrorMap } from '../../../../helpers/generateErrorMap';
 import { generateQuestions } from '../../../../helpers/generateQuestions';
 import { generateErrorSummary } from '../../../../helpers/generateErrorSummary';
-import { generateAddPriceTable } from '../../../../helpers/generateAddPriceTable';
+import { generateAddPriceTable } from '../../../../helpers/contextCreators/generateAddPriceTable';
 
 export const getContext = ({
   commonManifest,
